feat: initial query output tests
