+++ framework/material/BUILD
@@ -21,7 +21,8 @@ load("@io_bazel_rules_kotlin//kotlin:kotlin.bzl", "kt_jvm_library", "kt_jvm_test
 package(default_visibility = [
     "//module:__subpackages__",
     "//framework:__subpackages__",
-    "//:__pkg__"
+    "//:__pkg__",
+    "//test/integration:__pkg__"
 ])
 
 kt_jvm_library(
