+++ test/integration/BUILD
@@ -65,6 +65,8 @@ kt_jvm_test(
         "//state/app",
         "//state/connection",
         "//state/project",
+        "//state/common",
+        # External Maven Dependencies
         "@maven//:junit_junit",
         "@maven//:org_jetbrains_compose_foundation_foundation_desktop",
         "@maven//:org_jetbrains_compose_foundation_foundation_layout_desktop",
@@ -107,6 +109,7 @@ kt_jvm_test(
         "//state/connection",
         "//state/project",
         "//state/page",
+        # External Maven Dependencies
         "@maven//:junit_junit",
         "@maven//:org_jetbrains_compose_foundation_foundation_desktop",
         "@maven//:org_jetbrains_compose_foundation_foundation_layout_desktop",
@@ -148,6 +151,9 @@ kt_jvm_test(
         "//state/connection",
         "//state/project",
         "//state/page",
+        "//state/common",
+        "//framework/material",
+        # External Maven Dependencies
         "@maven//:junit_junit",
         "@maven//:org_jetbrains_compose_foundation_foundation_desktop",
         "@maven//:org_jetbrains_compose_foundation_foundation_layout_desktop",
@@ -189,6 +195,7 @@ kt_jvm_test(
         "//state/connection",
         "//state/project",
         "//state/page",
+        # External Maven Dependencies
         "@maven//:junit_junit",
         "@maven//:org_jetbrains_compose_foundation_foundation_desktop",
         "@maven//:org_jetbrains_compose_foundation_foundation_layout_desktop",
@@ -230,6 +237,7 @@ kt_jvm_test(
         "//state/connection",
         "//state/project",
         "//state/page",
+        # External Maven Dependencies
         "@maven//:junit_junit",
         "@maven//:org_jetbrains_compose_foundation_foundation_desktop",
         "@maven//:org_jetbrains_compose_foundation_foundation_layout_desktop",
