+++ test/integration/QueryOutput.kt
@@ -22,9 +22,13 @@
 
 package com.vaticle.typedb.studio.test.integration
 
+import androidx.compose.ui.test.assertHasClickAction
 import androidx.compose.ui.test.junit4.ComposeContentTestRule
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
+import androidx.compose.ui.test.onRoot
+import androidx.compose.ui.test.performClick
+import androidx.compose.ui.test.printToString
 import com.vaticle.typedb.client.TypeDB
 import com.vaticle.typedb.client.api.TypeDBOptions
 import com.vaticle.typedb.client.api.TypeDBSession
@@ -32,6 +36,7 @@ import com.vaticle.typedb.client.api.TypeDBTransaction
 import com.vaticle.typedb.studio.test.integration.runComposeRule
 import com.vaticle.typedb.studio.Studio
 import com.vaticle.typedb.studio.framework.common.WindowContext
+import com.vaticle.typedb.studio.framework.material.Icon
 import com.vaticle.typedb.studio.state.StudioState
 import com.vaticle.typedb.studio.state.project.FileState
 import com.vaticle.typedb.studio.state.project.PathState
@@ -46,6 +51,7 @@ import org.junit.Test
 import java.io.File
 import java.nio.charset.StandardCharsets
 import java.nio.file.Files
+import java.nio.file.Path
 import java.nio.file.Paths
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
@@ -54,22 +60,16 @@ class QueryOutput {
     companion object {
         private const val DB_ADDRESS = "localhost:1729"
         private const val DB_NAME = "github"
+
+        val SAMPLE_DATA_PATH = "test/data/sample_file_structure"
+
+        private val SAVE_ICON_STRING = Icon.Code.FLOPPY_DISK.unicode
+        private val PLUS_ICON_STRING = Icon.Code.PLUS.unicode
     }
 
     @get:Rule
     val composeRule = createComposeRule()
 
-    @Test
-    fun `Save File`() {
-        runComposeRule(composeRule) {
-            setContent {
-                Studio.MainWindowContent(WindowContext(1000, 1000, 0, 0))
-            }
-            composeRule.waitForIdle()
-        }
-    }
-
-    @Ignore
     @Test
     fun `Make a New File and Save It`() {
         runComposeRule(composeRule) {
@@ -77,8 +77,30 @@ class QueryOutput {
                 Studio.MainWindowContent(WindowContext(1000, 1000, 0, 0))
             }
             composeRule.waitForIdle()
+
+            // We have to open a project to enable the '+' to create a new file.
+            openProject(composeRule, SAMPLE_DATA_PATH)
+
+            composeRule.onNodeWithText(PLUS_ICON_STRING).performClick()
+
+            composeRule.waitForIdle()
+            delay(500)
+
+            composeRule.onNodeWithText("Untitled1.tql *").assertExists()
+
+            // Clicking this takes us to a window to choose where to save - so we just assert that it exists and
+            // can be clicked.
+            composeRule.onNodeWithText(SAVE_ICON_STRING).assertExists().assertHasClickAction()
+
+            StudioState.pages.active!!.initiateSave()
+            composeRule.waitForIdle()
+            delay(500)
+
+            assertTrue(File("Untitled1.tql").exists())
         }
     }
+
+
 
     @Ignore
     @Test
